rockchip-libmali: only fix up /lib/firmware if present
Only valhall platforms include a firmware blob, so fix
the do_install function to check if it's present before
trying to fix up its location for usrmerge builds.

+++ recipes-graphics/rockchip-libmali/rockchip-libmali_git.bb
@@ -75,7 +75,7 @@ do_install:append () {
 			${D}${includedir}/EGL/eglplatform.h
 	fi
 	# meson.build file hard-codes firmware installation to /lib/firmware
-	if [ "${nonarch_base_libdir}" != "/lib" ]; then
+	if [ -d "${D}/lib" -a "${nonarch_base_libdir}" != "/lib" ]; then
 	    mv ${D}/lib/firmware ${D}${nonarch_base_libdir}/
 	    rmdir ${D}/lib
 	fi
